Time: 0 ms (100%), Space: 7.8 MB (17.55%) - LeetHub

+++ 0367-valid-perfect-square/0367-valid-perfect-square.cpp
@@ -1,9 +1,9 @@
 class Solution {
 public:
-    bool isPerfectSquare(unsigned long long num) {
-        unsigned long long start = 1, end = num;
+    bool isPerfectSquare(long num) {
+        long start = 1, end = num;
         while(start <= end){
-            unsigned long long mid = start + (end - start) / 2;
+            long mid = start + (end - start) / 2;
             if(mid * mid == num)
                 return true;
             if(mid * mid > num){
